\name{dirac1}
\alias{dirac1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ dirac1 }
\description{
Indicator function of the set A, where A = [u,infinity)}
\usage{
dirac1(u, x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{u}{ real value }
  \item{x}{ real value }
}


\value{
  dirac1(u,x)=1 if x>=u and dirac1(u,x)=0 in other case.
}
\author{Veronica A. Gonzalez-Lopez }
\seealso{  \code{\link{dirac2}},  \code{\link{diracS1}},  \code{\link{diracS2}} }
\keyword{arith}
