\name{diracS1}
\alias{diracS1}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{ diracS1}
\description{
  Indicator function of the set A, where A = (-infinity,u)}
\usage{
diracS1(u, x)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{u}{ real value }
  \item{x}{ real value }
}

\value{
 diracS1(u,x)=1 if x<u and diracS1(u,x)=0 in other case.}
\author{ Veronica A. Gonzalez-Lopez }
\seealso{  \code{\link{dirac1}},  \code{\link{dirac2}},  \code{\link{diracS2}} }
\keyword{arith}